Clean-up mocked version of napari-skimage-regionprops (#497)
