--- tests/data/napari_workflows/mock_package/MANIFEST.in
@@ -1,6 +1,0 @@
-include requirements.txt
-
-recursive-exclude * __pycache__
-recursive-exclude demo *
-recursive-exclude images *
-recursive-exclude * *.py[co]
