+++ tests/data/napari_workflows/mock_package/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+poetry run python -m build --wheel
+rm -r build
